+++ backend/src/main/java/com/stampcrush/backend/api/manager/cafe/ManagerCafeCouponSettingApiController.java
@@ -1,9 +1,9 @@
 package com.stampcrush.backend.api.manager.cafe;
 
-import com.stampcrush.backend.config.resolver.OwnerAuth;
 import com.stampcrush.backend.api.manager.cafe.request.CafeCouponSettingUpdateRequest;
 import com.stampcrush.backend.application.manager.cafe.ManagerCafeCouponSettingService;
 import com.stampcrush.backend.application.manager.cafe.dto.CafeCouponSettingDto;
+import com.stampcrush.backend.config.resolver.OwnerAuth;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -17,9 +17,10 @@ public class ManagerCafeCouponSettingApiController {
     private final ManagerCafeCouponSettingService managerCafeCouponSettingService;
 
     @PostMapping
-    public ResponseEntity<Void> updateCafeCouponSetting(OwnerAuth owner,
-                                                        @RequestParam("cafe-id") Long cafeId,
-                                                        @RequestBody @Valid CafeCouponSettingUpdateRequest request
+    public ResponseEntity<Void> updateCafeCouponSetting(
+            OwnerAuth owner,
+            @RequestParam("cafe-id") Long cafeId,
+            @RequestBody @Valid CafeCouponSettingUpdateRequest request
     ) {
         CafeCouponSettingDto cafeCouponSettingDto = request.toCouponSettingDto();
         managerCafeCouponSettingService.updateCafeCouponSetting(cafeId, cafeCouponSettingDto);
