+++ backend/src/test/java/com/stampcrush/Test.java
@@ -1,0 +1,9 @@
+package com.stampcrush;
+
+public class Test {
+
+//    @org.junit.jupiter.api.Test
+//    void test() {
+//        LocalTime of = LocalTime.of(25, 1, 6);
+//    }
+}
