+++ backend/src/test/java/com/stampcrush/backend/api/cafe/ManagerCafeFindApiControllerTest.java
@@ -1,10 +1,8 @@
 package com.stampcrush.backend.api.cafe;
 
-import com.stampcrush.backend.api.manager.cafe.ManagerCafeApiController;
-import com.stampcrush.backend.application.manager.cafe.ManagerCafeService;
+import com.stampcrush.backend.api.manager.cafe.ManagerCafeFindApiController;
 import com.stampcrush.backend.entity.user.Owner;
 import com.stampcrush.backend.repository.user.OwnerRepository;
-import com.stampcrush.backend.repository.user.RegisterCustomerRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,17 +20,14 @@ import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(ManagerCafeApiController.class)
-class ManagerCafeApiControllerTest {
+@WebMvcTest(ManagerCafeFindApiController.class)
+class ManagerCafeFindApiControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
-    @MockBean
-    private ManagerCafeService managerCafeService;
+
     @MockBean
     private OwnerRepository ownerRepository;
-    @MockBean
-    private RegisterCustomerRepository customerRepository;
 
     private Owner owner;
     private String basicAuthHeader;
